common/common_config/requirements.txt
build
python-dotenv